combinig menu with game
